Move utility error functions to Text.Pandoc.Shared
